Update composer.json
Package facebook/graph-sdk is abandoned, you should avoid using it. No replacement was suggested.

This is an unofficial version of Facebook's PHP SDK designed for PHP 7/8+. It is being maintained and tested against the newest PHP versions. You can use this in place of version 5.x of Facebook's deprecated facebook/graph-sdk package.